refactor(framework): group plugin and ui modules into subdirectories
Sub-group the framework layer per #424 finding #2: move plugin-loader,
plugin-registry and plugin-version into framework/plugins/, and help and
completion into framework/ui/. Command machinery stays at framework/ top.

Adjust runtime relative-path resolution for the deeper nesting:
- help.ts reads package.json three levels up (framework/ui -> repo root)
- plugin-loader defaultPluginsCandidateDirs adds one level so production
  resolves dist/default-plugins from dist/framework/plugins

The layering guard is unaffected: layerOf() classifies by first path
segment, so framework/plugins and framework/ui remain 'framework'.

Refs #424
