+++ src/framework/plugins/plugin-version.ts
