+++ tests/unit/000-setup.test.ts
@@ -7,7 +7,7 @@ import { existsSync, rmSync } from "node:fs";
 import { join } from "node:path";
 import { after, before } from "node:test";
 import { getUserDataDir } from "../../src/core/config.ts";
-import { clearLoadedPlugins } from "../../src/framework/plugin-loader.ts";
+import { clearLoadedPlugins } from "../../src/framework/plugins/plugin-loader.ts";
 
 // Store original environment
 const originalEnv = { ...process.env };
