+++ tests/unit/default-plugins-path.test.ts
@@ -1,12 +1,12 @@
 /**
  * Regression guard: default-plugins directory resolution (issue #414 fallout).
  *
- * The plugin loader lives at `src/framework/plugin-loader.ts` (dev) and
- * `dist/framework/plugin-loader.js` (production). It must resolve the
- * bundled default plugins relative to its own location:
+ * The plugin loader lives at `src/framework/plugins/plugin-loader.ts` (dev)
+ * and `dist/framework/plugins/plugin-loader.js` (production). It must resolve
+ * the bundled default plugins relative to its own location:
  *
- *   - dev  (`src/framework/`)  → `<repo>/default-plugins`      (two levels up)
- *   - prod (`dist/framework/`) → `<repo>/dist/default-plugins` (one level up)
+ *   - dev  (`src/framework/plugins/`)  → `<repo>/default-plugins`      (three levels up)
+ *   - prod (`dist/framework/plugins/`) → `<repo>/dist/default-plugins` (two levels up)
  *
  * The unit + integration suites always run from the TS sources (dev
  * layout), so a wrong *production* depth would never surface at test
@@ -18,11 +18,11 @@
 import assert from "node:assert";
 import { join } from "node:path";
 import { describe, test } from "node:test";
-import { defaultPluginsCandidateDirs } from "../../src/framework/plugin-loader.ts";
+import { defaultPluginsCandidateDirs } from "../../src/framework/plugins/plugin-loader.ts";
 
 describe("default-plugins path resolution", () => {
 	test("production layout resolves into dist/default-plugins", () => {
-		const distLoaderDir = join("/opt", "app", "dist", "framework");
+		const distLoaderDir = join("/opt", "app", "dist", "framework", "plugins");
 		const candidates = defaultPluginsCandidateDirs(distLoaderDir);
 
 		const expected = join("/opt", "app", "dist", "default-plugins");
@@ -41,7 +41,7 @@ describe("default-plugins path resolution", () => {
 	});
 
 	test("development layout resolves into <repo>/default-plugins", () => {
-		const srcLoaderDir = join("/opt", "app", "src", "framework");
+		const srcLoaderDir = join("/opt", "app", "src", "framework", "plugins");
 		const candidates = defaultPluginsCandidateDirs(srcLoaderDir);
 
 		const expected = join("/opt", "app", "default-plugins");
@@ -52,7 +52,7 @@ describe("default-plugins path resolution", () => {
 	});
 
 	test("returns production candidate before development candidate", () => {
-		const distLoaderDir = join("/opt", "app", "dist", "framework");
+		const distLoaderDir = join("/opt", "app", "dist", "framework", "plugins");
 		const candidates = defaultPluginsCandidateDirs(distLoaderDir);
 
 		assert.strictEqual(candidates.length, 2);
